Split BeamStatsMonitor package helpers
