+++ xtrack/monitors/beam_stats_monitor/__init__.py
@@ -1,0 +1,1 @@
+from .beam_stats_monitor import BeamStatsMonitor
